Refactor HttpServer
