+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/services/TransactionStatusTest.java
@@ -66,10 +66,10 @@ package com.radixdlt.integration.services;
 
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
-import com.radixdlt.api.data.TransactionStatus;
+import com.radixdlt.api.archive.transaction.TransactionStatus;
 import com.radixdlt.api.service.transactions.BerkeleyTransactionsByIdStore;
-import com.radixdlt.api.archive.transactions.TransactionStatusService;
-import com.radixdlt.api.archive.transactions.TransactionStatusServiceModule;
+import com.radixdlt.api.archive.transaction.TransactionStatusService;
+import com.radixdlt.api.archive.transaction.TransactionStatusServiceModule;
 import com.radixdlt.engine.RadixEngine;
 import com.radixdlt.environment.Environment;
 import com.radixdlt.environment.deterministic.MultiNodeDeterministicRunner;
@@ -120,7 +120,7 @@ import java.util.List;
 import java.util.function.Supplier;
 import java.util.stream.Collectors;
 
-import static com.radixdlt.api.data.TransactionStatus.TRANSACTION_NOT_FOUND;
+import static com.radixdlt.api.archive.transaction.TransactionStatus.TRANSACTION_NOT_FOUND;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.assertEquals;
