+++ radixdlt-core/radixdlt/src/integration/java/com/radixdlt/integration/staking/StakingUnstakingValidatorsTest.java
@@ -67,9 +67,9 @@ package com.radixdlt.integration.staking;
 import com.google.common.collect.ClassToInstanceMap;
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.Multibinder;
-import com.radixdlt.api.archive.accounts.BerkeleyAccountInfoStore;
+import com.radixdlt.api.archive.account.BerkeleyAccountInfoStore;
 import com.radixdlt.api.data.ValidatorUptime;
-import com.radixdlt.api.archive.validators.BerkeleyValidatorUptimeArchiveStore;
+import com.radixdlt.api.archive.validator.BerkeleyValidatorUptimeArchiveStore;
 import com.radixdlt.api.archive.tokens.BerkeleyResourceInfoStore;
 import com.radixdlt.application.validators.scrypt.ValidatorUpdateRakeConstraintScrypt;
 import com.radixdlt.consensus.LedgerProof;
