+++ radixdlt-core/radixdlt/src/main/java/com/radixdlt/api/archive/validator/ValidatorApiModule.java
@@ -1,10 +1,9 @@
-/* Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
- *
+/*
+ * Copyright 2021 Radix Publishing Ltd incorporated in Jersey (Channel Islands).
  * Licensed under the Radix License, Version 1.0 (the "License"); you may not use this
  * file except in compliance with the License. You may obtain a copy of the License at:
  *
  * radixfoundation.org/licenses/LICENSE-v1
- *
  * The Licensor hereby grants permission for the Canonical version of the Work to be
  * published, distributed and used under or by reference to the Licensor’s trademark
  * Radix ® and use of any unregistered trade names, logos or get-up.
@@ -62,19 +61,56 @@
  * permissions under this License.
  */
 
-package com.radixdlt.api.data;
+package com.radixdlt.api.archive.validator;
 
-import org.json.JSONObject;
+import com.google.inject.AbstractModule;
+import com.google.inject.Inject;
+import com.google.inject.Provider;
+import com.google.inject.Scopes;
+import com.google.inject.multibindings.MapBinder;
+import com.google.inject.multibindings.Multibinder;
+import com.radixdlt.api.util.Controller;
+import com.radixdlt.api.util.JsonRpcController;
+import com.radixdlt.api.util.JsonRpcHandler;
+import com.radixdlt.api.util.JsonRpcServer;
+import com.radixdlt.store.berkeley.BerkeleyAdditionalStore;
 
-public enum TransactionStatus {
-	PENDING,
-	CONFIRMED,
-	FAILED,
-	TRANSACTION_NOT_FOUND;
+import java.lang.annotation.Annotation;
+import java.util.Map;
 
-	public JSONObject asJson() {
-		var key = this == TRANSACTION_NOT_FOUND ? "failure" : "status";
+public class ValidatorApiModule extends AbstractModule {
+	private final Class<? extends Annotation> annotationType;
+	private final String path;
 
-		return new JSONObject().put(key, name());
+	public ValidatorApiModule(Class<? extends Annotation> annotationType, String path) {
+		this.annotationType = annotationType;
+		this.path = path;
+	}
+
+	@Override
+	public void configure() {
+		var binder = Multibinder.newSetBinder(binder(), BerkeleyAdditionalStore.class);
+		bind(ArchiveValidationHandler.class).in(Scopes.SINGLETON);
+		bind(BerkeleyValidatorUptimeArchiveStore.class).in(Scopes.SINGLETON);
+		binder.addBinding().to(BerkeleyValidatorUptimeArchiveStore.class);
+
+		MapBinder.newMapBinder(binder(), String.class, Controller.class, annotationType)
+			.addBinding(path)
+			.toProvider(ControllerProvider.class);
+	}
+
+
+	private static class ControllerProvider implements Provider<Controller> {
+		@Inject
+		private ArchiveValidationHandler handler;
+
+		@Override
+		public Controller get() {
+			var handlers = Map.<String, JsonRpcHandler>of(
+				"validators.get_next_epoch_set", handler::handleValidatorsGetNextEpochSet,
+				"validators.lookup_validator", handler::handleValidatorsLookupValidator
+			);
+			return new JsonRpcController(new JsonRpcServer(handlers));
+		}
 	}
 }
