+++ radixdlt-core/radixdlt/src/test/java/com/radixdlt/api/handler/ArchiveValidationHandlerTest.java
@@ -64,14 +64,14 @@
 package com.radixdlt.api.handler;
 
 import com.radixdlt.api.data.ValidatorUptime;
-import com.radixdlt.api.archive.validators.ArchiveValidationHandler;
+import com.radixdlt.api.archive.validator.ArchiveValidationHandler;
 import com.radixdlt.networks.Addressing;
 import com.radixdlt.networks.Network;
 import org.json.JSONObject;
 import org.junit.Test;
 
-import com.radixdlt.api.data.ValidatorInfoDetails;
-import com.radixdlt.api.archive.validators.ValidatorArchiveInfoService;
+import com.radixdlt.api.archive.validator.ValidatorInfoDetails;
+import com.radixdlt.api.archive.validator.ValidatorArchiveInfoService;
 import com.radixdlt.crypto.ECKeyPair;
 import com.radixdlt.crypto.ECPublicKey;
 import com.radixdlt.identifiers.REAddr;
